+++ ashley/tests/com/badlogic/ashley/core/AddRemoveSystemDuringIterationTest.java
@@ -31,12 +31,11 @@ public class AddRemoveSystemDuringIterationTest {
 
     @Test
     public void addSystemDuringIterationTest() {
-        engine = new Engine();
         EntitySystem system1 = new NamedSystem(1) {
         };
         EntitySystem system10 = new NamedSystem(10) {
         };
-        final EntitySystem system4 = new NamedSystem(4) {  //System added during update
+        final EntitySystem system4 = new NamedSystem(4) {  //System added during update by system 5
         };
         EntitySystem system5 = new NamedSystem(5) {
 
@@ -54,18 +53,51 @@ public class AddRemoveSystemDuringIterationTest {
         engine.addSystem(system10);
         engine.addSystem(system5);
 
-        engine.update(1); //it add system 45 add system 4 during iteration
+        engine.update(1); //system 5 add system 4 during iteration
 
         List<Integer> expected = Arrays.asList(1, 5, 10); //system 4 should not be called
-        Assert.assertArrayEquals(expected.toArray(), systemCallOrder.toArray()); // result 1, 5, 5, 10
+        Assert.assertArrayEquals(expected.toArray(), systemCallOrder.toArray());
 
         systemCallOrder.clear();
         engine.update(1);
         expected = Arrays.asList(1, 4, 5, 10); //second update ok
         Assert.assertArrayEquals(expected.toArray(), systemCallOrder.toArray());
-
     }
+    @Test
+    public void removeSystemDuringIterationTest() {
+        engine = new Engine();
+        EntitySystem system1 = new NamedSystem(1) {
+        };
+        EntitySystem system10 = new NamedSystem(10) {
+        };
+        final EntitySystem system8 = new NamedSystem(8) {  //System remove during update by system 5
+        };
+        EntitySystem system5 = new NamedSystem(5) {
 
+            @Override
+            public void update(float deltaTime) {
+                super.update(deltaTime);
+                if (!firstUpdateDone) {
+                    getEngine().removeSystem(system8);
+                    firstUpdateDone = true;
+                }
+            }
+        };
+
+        engine.addSystem(system1);
+        engine.addSystem(system10);
+        engine.addSystem(system5);
+
+        engine.update(1); //system 5 remove system 8 during iteration
+
+        List<Integer> expected = Arrays.asList(1, 5, 10); //system 8 should not be called
+        Assert.assertArrayEquals(expected.toArray(), systemCallOrder.toArray());
+
+        systemCallOrder.clear();
+        engine.update(1);
+        expected = Arrays.asList(1, 5, 10); //second update ok
+        Assert.assertArrayEquals(expected.toArray(), systemCallOrder.toArray());
+    }
     private class NamedSystem extends EntitySystem {
 
         public NamedSystem(int priority) {